clean up and added session storage
