+++ src/frontend/e2e/fixtures/userInfo.ts
