+++ src/frontend/e2e/setup/local.config.ts
