+++ src/frontend/e2e/setup/staging.config.ts
@@ -13,7 +13,8 @@ const config: PlaywrightTestConfig = {
   reporter: [ ['html', {open: 'never'} ] ],
   globalSetup: './global-setup',
   use: {
-    storageState: 'playwright/storage/state.json',
+    baseURL: "https://staging.czgenepi.org",
+    //storageState: 'playwright/storage/state.json',
     actionTimeout: 0,
     trace: 'on-first-retry',
     screenshot: "only-on-failure",
