+++ src/frontend/e2e/staticFiles/security.test.ts
