+++ src/frontend/package.json
@@ -15,13 +15,11 @@
     "type-check": "tsc --noEmit",
     "lint": "concurrently \"node_modules/.bin/eslint **/*.{ts,tsx} --quiet --fix\" \"node_modules/.bin/stylelint --fix '**/*.{js,ts,tsx,css}'\"  \"npm run type-check\"",
     "lint-ci": "concurrently \"node_modules/.bin/eslint **/*.{ts,tsx}\" \"node_modules/.bin/stylelint '**/*.{js,ts,tsx,css}'\"  \"npm run type-check\"",
-    "pl:staging:headed": "npx playwright test --headed -c ./playwright/setup/staging.config.ts",
-    "pl:local:headed": "npx playwright test --headed -c ./playwright/setup/staging.config.ts",
-    "pl:staging:headless": "npx playwright test --headed -c ./playwright/setup/staging.config.ts",
-    "pl:staging:debug": "npx playwright test --debug -c ./playwright/setup/staging.config.ts",
-    "pl:report": "npx playwright show-report",
-    "cy:staging:open": "npx cypress open  --config-file ./cypress/staging.config.ts",
-    "cy:staging:run": "npx cypress open  --config-file ./cypress/staging.config.ts"
+    "pl:staging:headed": "npx playwright test --headed -c ./e2e/setup/staging.config.ts",
+    "pl:local:headed": "npx playwright test --headed -c ./e2e/setup/staging.config.ts",
+    "pl:staging:headless": "npx playwright test --headed -c ./e2e/setup/staging.config.ts",
+    "pl:staging:debug": "npx playwright test --debug -c ./e2e/setup/staging.config.ts",
+    "pl:report": "npx playwright show-report"
   },
   "dependencies": {
     "@emotion/css": "^11.1.3",
